// web-application/src/types-openapi/api/models/ListSliderResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ListSliderResponse
 */
export interface ListSliderResponse {
    /**
     * 
     * @type {Date}
     * @memberof ListSliderResponse
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ListSliderResponse
     */
    updatedAt?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof ListSliderResponse
     */
    createdBy?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListSliderResponse
     */
    updatedBy?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ListSliderResponse
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof ListSliderResponse
     */
    title: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListSliderResponse
     */
    description: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListSliderResponse
     */
    imageUrl: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListSliderResponse
     */
    backLink?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListSliderResponse
     */
    status: string | null;
}

/**
 * Check if a given object implements the ListSliderResponse interface.
 */
export function instanceOfListSliderResponse(value: object): value is ListSliderResponse {
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('description' in value) || value['description'] === undefined) return false;
    if (!('imageUrl' in value) || value['imageUrl'] === undefined) return false;
    if (!('status' in value) || value['status'] === undefined) return false;
    return true;
}

export function ListSliderResponseFromJSON(json: any): ListSliderResponse {
    return ListSliderResponseFromJSONTyped(json, false);
}

export function ListSliderResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListSliderResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'updatedAt': json['updatedAt'] == null ? undefined : (new Date(json['updatedAt'])),
        'createdBy': json['createdBy'] == null ? undefined : json['createdBy'],
        'updatedBy': json['updatedBy'] == null ? undefined : json['updatedBy'],
        'id': json['id'] == null ? undefined : json['id'],
        'title': json['title'],
        'description': json['description'],
        'imageUrl': json['imageUrl'],
        'backLink': json['backLink'] == null ? undefined : json['backLink'],
        'status': json['status'],
    };
}

export function ListSliderResponseToJSON(json: any): ListSliderResponse {
    return ListSliderResponseToJSONTyped(json, false);
}

export function ListSliderResponseToJSONTyped(value?: ListSliderResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
        'updatedAt': value['updatedAt'] === null ? null : ((value['updatedAt'] as any)?.toISOString()),
        'createdBy': value['createdBy'],
        'updatedBy': value['updatedBy'],
        'id': value['id'],
        'title': value['title'],
        'description': value['description'],
        'imageUrl': value['imageUrl'],
        'backLink': value['backLink'],
        'status': value['status'],
    };
}